requirements.txt
fastapi
uvicorn
llama-index
pydantic
jinja2
llama-index-llms-gemini==0.1.10
google-generativeai==0.5.4
python-dotenv==1.0.1
pydantic[email]
psycopg2-binary
alembic